Fixed issues with clustering because clustering docs suck
