--- initializers/cluster_start.js
@@ -1,3 +1,0 @@
-module.exports = function(cluster) {
-  if (cluster.clusteringActive) cluster.init()
-}
