--- initializers/prometheus_server.js
@@ -1,3 +1,0 @@
-module.exports = (prometheus, cluster) => {
-  prometheus.init(cluster)
-}
