+++ lib/index.js
@@ -1,5 +1,10 @@
 require('./promise')
 
+const cluster = require('cluster')
+const os = require('os')
+const http = require('http')
+const promClient = require('prom-client')
+
 const EventEmitter = require('./event_emitter')
 const Errors = require('./errors/errors')
 const Container = require('./di/container')
@@ -27,6 +32,34 @@ process.on('unhandledRejection', function(reason, promise) {
   }
 })
 
+function metricsServerHandler(req, res) {
+  if (req.method === 'GET' && req.url === '/metrics') {
+    res.writeHead(200, { 'Content-Type': 'text/plain' })
+    res.end(promClient.register.metrics())
+  } else {
+    res.writeHead(404).end()
+  }
+}
+
+function aggregatedMetricsServerHandler(req, res) {
+  if (req.method === 'GET' && req.url === '/metrics') {
+    this.aggr.clusterMetrics((err, metrics) => {
+      if (err) {
+        Q.log.warn(err)
+        res.writeHead(500, 'Content-Type', 'text/plain')
+        res.end(err.message)
+      } else {
+        res.writeHead(200, 'Content-Type', this.aggr.contentType)
+        res.end(metrics)
+      }
+    })
+    res.writeHead(200, { 'Content-Type': 'text/plain' })
+    res.end(promClient.register.metrics())
+  } else {
+    res.writeHead(404).end()
+  }
+}
+
 async function boot(opts) {
   opts = opts || {}
 
@@ -38,6 +71,35 @@ async function boot(opts) {
   Q.config.loadFrom(Q.plugins.map(_ => _.pluginDir))
 
   Q.log = await container.getAsync('log')
+  const cmd = Q.app.getAppCommand()
+
+  // start workers
+  const clustering = (
+    Q.config.get('quadro.clustering', false) &&
+    ['test', 'repl', 'task'].indexOf(cmd) < 0
+  )
+
+  if (cluster.isMaster && clustering) {
+    for (let i = 0; i < os.cpus(); ++i) cluster.fork()
+    cluster.on('exit', (worker, code, signal) => {
+      Q.log.info(`Worker ${worker.process.pid} exited with code ${code} from signal ${signal}.`)
+      this.metrics.workerCount.dec()
+      process.exit(code)
+    })
+  }
+
+  // start prometheus server
+  if (clustering && cluster.isMaster) {
+    const aggr = new promClient.AggregatorRegistry()
+    http.createServer(aggregatedMetricsServerHandler.bind({aggr}))
+      .listen(Q.config.get('prometheus.aggregatorPort', 9230))
+  } else if (!clustering) {
+    http.createServer(metricsServerHandler)
+      .listen(Q.config.get('prometheus.port', 9230))
+  } // else, there is clustering, and this is not the master, so do nothing
+
+  // prevent master from doing work
+  if (clustering && cluster.isMaster) return
 
   await container.run(require('./services_loader'))
   await container.run(require('./models_loader'))
@@ -49,7 +111,6 @@ async function boot(opts) {
 
   await Q.app.emit('loaded', Q.app)
 
-  let cmd = Q.app.getAppCommand()
   if (cmd === 'test') return runTests()
   if (cmd === 'repl') return runREPL()
   if (cmd === 'task') return runTask()
